+++ uplink/arguments.py
@@ -24,6 +24,7 @@ __all__ = [
     "Body",
     "Url",
     "Timeout",
+    "Context",
 ]
 
 
@@ -119,6 +120,9 @@ class ArgumentAnnotationHandlerBuilder(interfaces.AnnotationHandlerBuilder):
 
     def is_done(self):
         return self.remaining_args_count == 0
+
+    def copy(self):
+        return self
 
     @property
     def _types(self):
@@ -680,7 +684,7 @@ class Url(ArgumentAnnotation):
 
 class Timeout(FuncDecoratorMixin, ArgumentAnnotation):
     """
-    Pass a timeout as a method argument at runtime.
+    Passes a timeout as a method argument at runtime.
 
     While :py:class:`uplink.timeout` attaches static timeout to all requests
     sent from a consumer method, this class turns a method argument into a
@@ -693,7 +697,6 @@ class Timeout(FuncDecoratorMixin, ArgumentAnnotation):
             def get_posts(self, timeout: Timeout() = 60):
                 \"""Fetch all posts for the current users giving up after given
                 number of seconds.\"""
-
     """
 
     @property
@@ -708,3 +711,78 @@ class Timeout(FuncDecoratorMixin, ArgumentAnnotation):
     def _modify_request(self, request_builder, value):
         """Modifies request timeout."""
         request_builder.info["timeout"] = value
+
+
+class Context(FuncDecoratorMixin, NamedArgument):
+    """
+    Defines a name-value pair that is accessible to middleware at
+    runtime.
+
+    Request middleware can leverage this annotation to give users
+    control over the middleware's behavior.
+
+    Example:
+        Consider a custom decorator :obj:`@cache` (this would be a
+        subclass of :class:`uplink.decorators.MethodAnnotation`):
+
+        .. code-block:: python
+
+            @cache(hours=3)
+            @get("users/user_id")
+            def get_user(self, user_id)
+                \"""Retrieves a single user.\"""
+
+        As its name suggests, the :obj:`@cache` decorator enables
+        caching server responses so that, once a request is cached,
+        subsequent identical requests can be served by the cache
+        provider rather than adding load to the upstream service.
+
+        Importantly, the writers of the :obj:`@cache` decorators can
+        allow users to pass their own cache provider implementation
+        through an argument annotated with :class:`Context <uplink.Context>`:
+
+        .. code-block:: python
+
+            @cache(hours=3)
+            @get("users/user_id")
+            def get_user(self, user_id, cache_provider: Context)
+                \"""Retrieves a single user.\"""
+
+        To add a name-value pair to the context of any request made from
+        a :class:`Consumer <uplink.Consumer>` instance, you
+        can use the :attr:`Consumer.session.context
+        <uplink.session.Session.context>` property. Alternatively, you
+        can annotate a constructor argument of a :class:`Consumer
+        <uplink.Consumer>` subclass with :class:`Context
+        <uplink.Context>`, as explained
+        :ref:`here <annotating constructor arguments>`.
+    """
+
+    @property
+    def converter_key(self):
+        """Do not convert passed argument."""
+        return keys.Identity()
+
+    def _modify_request(self, request_builder, value):
+        """Sets the name-value pair in the context."""
+        request_builder.context[self.name] = value
+
+
+class ContextMap(FuncDecoratorMixin, ArgumentAnnotation):
+    """
+    Defines a mapping of name-value pairs that are accessible to
+    middleware at runtime.
+    """
+
+    @property
+    def converter_key(self):
+        """Do not convert passed argument."""
+        return keys.Identity()
+
+    def _modify_request(self, request_builder, value):
+        """Updates the context with the given name-value pairs."""
+        if not isinstance(value, collections.Mapping):
+            raise TypeError(
+                "ContextMap requires a mapping; got %s instead.", type(value)
+            )
+        request_builder.context.update(value)