+++ uplink/commands.py
@@ -34,61 +34,15 @@ class MissingUriVariables(exceptions.InvalidRequestDefinition):
         self.message = self.message % (uri, "', '".join(remaining_variables))
 
 
-class HttpMethodFactory(object):
-    def __init__(self, method):
-        self._method = method
-
-    def __call__(self, uri=None, args=()):
-        if callable(uri) and not args:
-            return HttpMethod(self._method)(uri)
-        else:
-            return HttpMethod(self._method, uri, args)
-
-
-class HttpMethod(object):
-    @staticmethod
-    def _add_args(obj):
-        return obj
-
-    def __init__(self, method, uri=None, args=None):
-        self._method = method
-        self._uri = uri
-
-        # Register argument annotations
-        if args:
-            is_map = isinstance(args, collections.Mapping)
-            args, kwargs = ((), args) if is_map else (args, {})
-            self._add_args = decorators.args(*args, **kwargs)
-
-    def __call__(self, func):
-        spec = utils.get_arg_spec(func)
-        arg_handler = arguments.ArgumentAnnotationHandlerBuilder(
-            func, spec.args
-        )
-        builder = RequestDefinitionBuilder(
-            self._method,
-            URIDefinitionBuilder(self._uri),
-            arg_handler,
-            decorators.MethodAnnotationHandlerBuilder(),
-        )
-
-        # Need to add the annotations after constructing the request
-        # definition builder so it has a chance to attach its listener.
-        arg_handler.set_annotations(spec.annotations)
-
-        # Use return value type hint as expected return type
-        if spec.return_annotation is not None:
-            builder = returns.schema(spec.return_annotation)(builder)
-        functools.update_wrapper(builder, func)
-        builder = self._add_args(builder)
-        return builder
-
-
 class URIDefinitionBuilder(interfaces.UriDefinitionBuilder):
     def __init__(self, uri):
         self._uri = uri
         self._is_dynamic = False
         self._uri_variables = set()
+
+    @property
+    def template(self):
+        return self._uri
 
     @property
     def is_static(self):
@@ -132,6 +86,7 @@ class RequestDefinitionBuilder(interfaces.RequestDefinitionBuilder):
         self._uri = uri
         self._argument_handler_builder = argument_handler_builder
         self._method_handler_builder = method_handler_builder
+        self._return_type = None
 
         self._argument_handler_builder.listener = self._notify
         self._method_handler_builder.listener = self._notify
@@ -155,6 +110,106 @@ class RequestDefinitionBuilder(interfaces.RequestDefinitionBuilder):
     def method_handler_builder(self):
         return self._method_handler_builder
 
+    @property
+    def return_type(self):
+        return self._return_type
+
+    @return_type.setter
+    def return_type(self, return_type):
+        self._return_type = return_type
+
+    def __call__(self, uri=None, args=()):
+        """
+        Applies the decorators, HTTP method, and optionally the URI
+        of this consumer method to the decorated method.
+
+        This makes the request definition reusable and can help
+        minimize duplication across similar consumer methods.
+
+        Examples:
+
+            Define request templates:
+
+            .. code-block:: python
+
+                from uplink import Consumer, get, json, returns
+
+                @returns.json
+                @json
+                @get
+                def get_json():
+                    \"""GET request that consumes and produces JSON.\"""
+
+                class GitHub(Consumer):
+                    @get_json("/users/{user}")
+                    def get_user(self, user):
+                         \"""Fetches a specific GitHub user.\"""
+
+            Remove duplication across definitions of similar consumer
+            methods, whether or not the methods are defined in the same
+            class:
+
+            .. code-block:: python
+
+                from uplink import Consumer, get, params, timeout
+
+                class GitHub(Consumer):
+                    @timeout(10)
+                    @get("/users/{user}/repos")
+                    def get_user_repos(self, user):
+                        \"""Retrieves the repos that the user owns.\"""
+
+                    # Extends the above method to define a variant:
+                    @params(type="member")
+                    @get_user_repos
+                    def get_repos_for_collaborator(self, user):
+                        \"""
+                        Retrieves the repos for which the given user is
+                        a collaborator.
+                        \"""
+
+                class EnhancedGitHub(Github):
+                    # Updates the return type of an inherited method.
+                    @GitHub.get_user_repos
+                    def get_user_repos(self, user) -> List[Repo]:
+                        \"""Retrieves the repos that the user owns.\"""
+
+        Args:
+            uri (str, optional): the request's relative path
+            args: a list or mapping of function annotations (e.g.
+                  :class:`uplink.Path`) corresponding to the decorated
+                  function's arguments
+        """
+        return self.extend(uri, args)
+
+    def extend(self, uri=None, args=()):
+        factory = HttpMethodFactory(
+            method=self.method, request_definition_builder_factory=self._extend
+        )
+
+        if callable(uri):
+            return factory(self.uri.template, args)(uri)
+        else:
+            uri = self.uri.template if uri is None else uri
+            return factory(uri, args)
+
+    def _extend(self, method, uri, arg_handler, _):
+        builder = RequestDefinitionBuilder(
+            method, uri, arg_handler, self.method_handler_builder.copy()
+        )
+        builder.return_type = self.return_type
+        return builder
+
+    def copy(self):
+        builder = RequestDefinitionBuilder(
+            self._method,
+            self._uri,
+            self._argument_handler_builder.copy(),
+            self._method_handler_builder.copy(),
+        )
+        builder.return_type = self.return_type
+        return builder
+
     def _auto_fill_remaining_arguments(self):
         uri_vars = set(self.uri.remaining_variables)
         missing = list(self.argument_handler_builder.missing_arguments)
@@ -176,14 +231,21 @@ class RequestDefinitionBuilder(interfaces.RequestDefinitionBuilder):
         method_handler = self._method_handler_builder.build()
         uri = self._uri.build()
         return RequestDefinition(
-            self._method, uri, argument_handler, method_handler
+            self._method,
+            uri,
+            self._return_type,
+            argument_handler,
+            method_handler,
         )
 
 
 class RequestDefinition(interfaces.RequestDefinition):
-    def __init__(self, method, uri, argument_handler, method_handler):
+    def __init__(
+        self, method, uri, return_type, argument_handler, method_handler
+    ):
         self._method = method
         self._uri = uri
+        self._return_type = return_type
         self._argument_handler = argument_handler
         self._method_handler = method_handler
 
@@ -201,11 +263,76 @@ class RequestDefinition(interfaces.RequestDefinition):
     def define_request(self, request_builder, func_args, func_kwargs):
         request_builder.method = self._method
         request_builder.url = utils.URIBuilder(self._uri)
+        request_builder.return_type = self._return_type
         self._argument_handler.handle_call(
             request_builder, func_args, func_kwargs
         )
         self._method_handler.handle_builder(request_builder)
         request_builder.url = request_builder.url.build()
+
+
+class HttpMethodFactory(object):
+    def __init__(
+        self,
+        method,
+        request_definition_builder_factory=RequestDefinitionBuilder,
+    ):
+        self._method = method
+        self._request_definition_builder_factory = (
+            request_definition_builder_factory
+        )
+
+    def __call__(self, uri=None, args=()):
+        if callable(uri) and not args:
+            return HttpMethod(self._method)(
+                uri, self._request_definition_builder_factory
+            )
+        else:
+            return functools.partial(
+                HttpMethod(self._method, uri, args),
+                request_definition_builder_factory=self._request_definition_builder_factory,
+            )
+
+
+class HttpMethod(object):
+    @staticmethod
+    def _add_args(obj):
+        return obj
+
+    def __init__(self, method, uri=None, args=None):
+        self._method = method
+        self._uri = uri
+
+        # Register argument annotations
+        if args:
+            is_map = isinstance(args, collections.Mapping)
+            args, kwargs = ((), args) if is_map else (args, {})
+            self._add_args = decorators.args(*args, **kwargs)
+
+    def __call__(
+        self, func, request_definition_builder_factory=RequestDefinitionBuilder
+    ):
+        spec = utils.get_arg_spec(func)
+        arg_handler = arguments.ArgumentAnnotationHandlerBuilder(
+            func, spec.args
+        )
+        builder = request_definition_builder_factory(
+            self._method,
+            URIDefinitionBuilder(self._uri),
+            arg_handler,
+            decorators.MethodAnnotationHandlerBuilder(),
+        )
+
+        # Need to add the annotations after constructing the request
+        # definition builder so it has a chance to attach its listener.
+        arg_handler.set_annotations(spec.annotations)
+
+        # Use return value type hint as expected return type
+        if spec.return_annotation is not None:
+            builder = returns.schema(spec.return_annotation)(builder)
+        functools.update_wrapper(builder, func)
+        builder = self._add_args(builder)
+        return builder
 
 
 get = HttpMethodFactory("GET").__call__