+++ uplink/retry/retry.py
@@ -12,28 +12,37 @@ __all__ = ["retry"]
 
 
 class RetryTemplate(RequestTemplate):
-    def __init__(self, back_off_iterator, retry_condition):
-        self._back_off_iterator = back_off_iterator
+    def __init__(self, backoff, retry_condition):
+        self._backoff = backoff
+        self._backoff_iterator = None
         self._condition = retry_condition
+        self._reset()
 
     def _next_delay(self):
         try:
-            delay = next(self._back_off_iterator)
+            delay = next(self._backoff_iterator)
         except StopIteration:
             # Fallback to the default behavior
             pass
         else:
             return transitions.sleep(delay)
 
+    def _reset(self):
+        self._backoff_iterator = self._backoff()
+
     def after_response(self, request, response):
         if self._condition.should_retry_after_response(response):
             return self._next_delay()
+        else:
+            self._reset()
 
     def after_exception(self, request, exc_type, exc_val, exc_tb):
         if self._condition.should_retry_after_exception(
             exc_type, exc_val, exc_tb
         ):
             return self._next_delay()
+        else:
+            self._reset()
 
 
 # noinspection PyPep8Naming
@@ -52,6 +61,21 @@ class retry(decorators.MethodAnnotation):
     `capped exponential backoff and jitter <https://amzn.to/2xc2nK2>`_,
     which should benefit performance with remote services under high
     contention.
+
+    .. note::
+
+        Response and error handlers (see :ref:`here <custom response
+        handler>`) are invoked after the retry condition breaks or all
+        retry attempts are exhausted, whatever comes first. These
+        handlers will receive the first response/exception that triggers
+        the retry's ``stop`` condition or doesn't match its ``when``
+        filter.
+
+        In other words, responses or exceptions that match
+        the retry condition (e.g., retry when status code is 5xx) are
+        not subject to response or error handlers as long as the request
+        doesn't break the retry's stop condition (e.g., stop retrying
+        after 5 attempts).
 
     Args:
         when (optional): A predicate that determines when a retry
@@ -83,7 +107,7 @@ class retry(decorators.MethodAnnotation):
         elif max_attempts is not None:
             self._stop = stop_mod.after_attempt(max_attempts)
         else:
-            self._stop = stop_mod.DISABLE
+            self._stop = stop_mod.NEVER
 
         self._predicate = when
 
@@ -110,7 +134,7 @@ class retry(decorators.MethodAnnotation):
 
     def _create_template(self, request_builder):
         return RetryTemplate(
-            self._backoff_iterator(), self._predicate(request_builder)
+            self._backoff_iterator, self._predicate(request_builder)
         )
 
     def _backoff_iterator(self):